LayoutTests/fast/js/kde/encode_decode_uri-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS checkWithFunctions(encodeURI,decodeURI) is true
PASS checkWithFunctions(encodeURIComponent,decodeURIComponent) is true
PASS successfullyParsed is true

TEST COMPLETE
